+++ atlas.com/npc/data/consumable/model.go
@@ -36,7 +36,7 @@ type Model struct {
 	id              uint32
 	tradeBlock      bool
 	price           uint32
-	unitPrice       uint32
+	unitPrice       float64
 	slotMax         uint32
 	timeLimited     bool
 	notSale         bool
@@ -174,7 +174,7 @@ func (m Model) CursedRate() uint32 {
 	return m.cursed
 }
 
-func (m Model) UnitPrice() uint32 {
+func (m Model) UnitPrice() float64 {
 	return m.unitPrice
 }
