+++ atlas.com/npc/data/etc/rest.go
@@ -7,7 +7,7 @@ import (
 type RestModel struct {
 	Id        uint32 `json:"-"`
 	Price     uint32 `json:"price"`
-	UnitPrice uint32 `json:"unitPrice"`
+	UnitPrice float64 `json:"unitPrice"`
 	SlotMax   uint32 `json:"slotMax"`
 }
